simplify the AndroidMakersStore API
