+++ store/src/main/kotlin/fr/androidmakers/store/model/Room.kt
@@ -1,4 +1,3 @@
 package fr.androidmakers.store.model
 
-
-data class Room(val name: String = "")
+data class Room(val id: String = "", val name: String = "")
